feat: enable panicnil=0 as it is since go1.21
